refactor: rename extensions/ to extension-core/
Rename packages/otter-agent/src/extensions/ to extension-core/ to resolve
ambiguity with the interface implementation naming convention.

The codebase uses a consistent pattern where interfaces/ defines pluggable
interfaces and sibling folders contain built-in implementations
(environments/, auth-storages/, session-managers/, ui-providers/). The
extensions/ folder broke this pattern — it contained extension system
infrastructure (ExtensionsAPI, EventBus, ExtensionRunner, event types,
config validation), not built-in extension implementations.

Renaming to extension-core/ makes the folder's purpose clear and
preserves extensions/ for future built-in extensions.

Closes #69
