+++ packages/otter-agent/src/extension-core/event-bus.ts
