+++ packages/otter-agent/src/extension-core/validate-extension-config.test.ts
