+++ packages/otter-agent/src/session/agent-session.ts
@@ -11,10 +11,10 @@ import type {
 	ThinkingLevel,
 } from "@mariozechner/pi-agent-core";
 import type { Api, ImageContent, Model } from "@mariozechner/pi-ai";
-import type { CompactOptions } from "../extensions/context.js";
-import { ExtensionRunner } from "../extensions/extension-runner.js";
-import type { ExtensionRunnerActions } from "../extensions/extension-runner.js";
-import type { Extension } from "../extensions/extension.js";
+import type { CompactOptions } from "../extension-core/context.js";
+import { ExtensionRunner } from "../extension-core/extension-runner.js";
+import type { ExtensionRunnerActions } from "../extension-core/extension-runner.js";
+import type { Extension } from "../extension-core/extension.js";
 import type { AgentEnvironment } from "../interfaces/agent-environment.js";
 import type { AuthStorage } from "../interfaces/auth-storage.js";
 import type { EntryId, SessionManager } from "../interfaces/session-manager.js";
